works with c flags

+++ src/ClientInfo.hpp
@@ -30,7 +30,7 @@ class Request;
 		time_t 					last_received;
 		std::string				response;
 		
-		ssize_t 				total_bytes_sent;
+		size_t 					total_bytes_sent;
 		Request					*request_obj;
 		bool					is_receiving;
 
@@ -41,7 +41,7 @@ class Request;
 
 		bool					is_reading;
 		size_t					total_bytes_read;
-		off_t					file_size;
+		size_t					file_size;
 		
 		bool					force_drop_connection;
